refactor(UIKIT-669,form,FormTextField): Вынес FormTextFieldValue в отдельный файл (#554)
