+++ packages/form/src/FormTextField/FormTextField.stories.tsx
@@ -3,8 +3,7 @@ import { Story } from '@storybook/react';
 import { useForm } from '../hooks';
 import { FormStoryContainer } from '../docs';
 import { FormSubmitButton } from '../FormSubmitButton';
-
-import { FormTextField, FormTextFieldValue } from './FormTextField';
+import { FormTextField, FormTextFieldValue } from '../FormTextField';
 
 export default {
   title: 'Form/FormTextField',
