+++ packages/form/src/FormTextField/index.ts
@@ -1,1 +1,3 @@
 export * from './FormTextField';
+
+export * from './types';
