+++ packages/form/src/FormTextField/types.ts
@@ -1,0 +1,4 @@
+/**
+ * @description Тип значения, которое сетится в state формы
+ */
+export type FormTextFieldValue = string;
